requirements.txt
Django==1.11.13
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Markdown==2.6.11
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.4
